refactor: 구 tokens 폴더 해체
다음 커밋에서 tokensV2 폴더를 tokens로 바꾸기 위해 구 tokens 폴더 해체

-  zIndex·animation 토큰은 구버전/신버전이 나뉘는 것이 아니라 앱 전체(v2 컴포넌트 포함)가 쓰는 유일본이라, 기존 zIndex, animation 파일 전체를 tokensV2 폴더로 위치만 이동
- 구 폰트 값(tokens/font.css.ts)은 신 토큰과 자간·행간 값이 달라 지금 전환하면 화면이 변하므로, fontLegacy.css.ts로 개명해 보존 (마지막 커밋에서 전환or유지 결정 예정)
