+++ src/pages/generate/v2/pages/result/components/list/imgSection/GeneratedImg.css.ts
@@ -2,7 +2,7 @@ import { globalStyle, style } from '@vanilla-extract/css';
 import { recipe } from '@vanilla-extract/recipes';
 
 import { fontStyle } from '@styles/fontStyle';
-import { animationTokens } from '@styles/tokens/animation.css';
+import { animationTokens } from '@styles/tokensV2/animation.css';
 import { colorVars } from '@styles/tokensV2/color.css';
 
 export const container = style({
