+++ src/shared/components/v2/menuTab/MenuTab.css.ts
@@ -1,11 +1,11 @@
 import { style } from '@vanilla-extract/css';
 import { recipe } from '@vanilla-extract/recipes';
 
-import { zIndex } from '@styles/tokens/zIndex';
 import { colorVars } from '@styles/tokensV2/color.css';
 import { fontVars } from '@styles/tokensV2/font.css';
 import { pressInteraction } from '@styles/tokensV2/interaction/presets';
 import { unitVars } from '@styles/tokensV2/unit.css';
+import { zIndex } from '@styles/tokensV2/zIndex';
 
 export const menuTabBar = recipe({
   base: {
