+++ src/shared/components/v2/popup/Popup.css.ts
@@ -6,8 +6,8 @@ import {
   popupFadeInInteraction,
   popupFadeInOpacityInteraction,
 } from '@styles/tokensV2/interaction/presets';
+import { zIndex } from '@styles/tokensV2/zIndex';
 
-import { zIndex } from '@/shared/styles/tokens/zIndex';
 import { fontVars } from '@/shared/styles/tokensV2/font.css';
 import { unitVars } from '@/shared/styles/tokensV2/unit.css';
 const motionHidden = {
