+++ src/shared/styles/tokensV2/zIndex.ts
@@ -1,14 +1,11 @@
 export const zIndex = {
-  base: 1,
   illustration: 2,
   text: 10,
-  blurButton: 40,
   button: 50,
   sticky: 100,
   navigation: 200,
   backdrop: 250,
   modal: 300,
-  toast: 400,
   navBar: 500,
   sheet: 550,
   popup: 600,
